fix: Prevent horizontal shifting in properties grid
