+++ src/AccessibilityInsights.SharedUx/Controls/PropertyInfoControl.xaml
@@ -18,7 +18,7 @@
     </UserControl.Resources>
     <Grid>
         <ScrollViewer HorizontalScrollBarVisibility="Auto" VerticalScrollBarVisibility="Auto" HorizontalAlignment="Stretch" >
-            <Grid >
+            <Grid RequestBringIntoView="Grid_RequestBringIntoView">
                 <Grid.RowDefinitions>
                     <RowDefinition Height="Auto"/>
                     <RowDefinition Height="*"/>
